refactor: split connectivity propagation for wave 5 slice B (#4125)
Extract road-rule BFS and port-expansion/town-closure passes into
dedicated libraries; keep connectedTilesForPlayer orchestration thin.

Refs #4125
