+++ packages/colonizethis_world/lib/src/world/connectivity_propagation.dart
@@ -1,11 +1,9 @@
-/// Capital-connectivity propagation core (Road rule, sea-path port wiring,
+/// Capital-connectivity propagation orchestration (Road rule, sea-path port wiring,
 /// blockade pruning, Town rule closure). Standalone library (extracted from the
 /// former `connectivity_resolver.dart` `part` chain, Refs #3544 Step 3) so the
 /// propagation algorithm can be understood, imported, and unit-tested without
-/// pulling in the resolver. Public surface is [connectedTilesForPlayer]; the
-/// rest is library-private. None of these are surfaced through the package
-/// barrel (they stay package-internal, consumed only by
-/// `connectivity_resolver.dart`).
+/// pulling in the resolver. Public surface is [connectedTilesForPlayer]; road BFS
+/// and port-expansion passes live in sibling libraries (Refs #4125 slice B).
 library;
 
 import 'dart:collection';
@@ -14,49 +12,10 @@ import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 
 import 'connectivity_metrics.dart';
+import 'connectivity_port_expansion.dart';
 import 'connectivity_result.dart';
+import 'connectivity_road_propagation.dart';
 import 'connectivity_tile_helpers.dart';
-import 'port_seaboard_registry_key.dart';
-import 'topology_helpers.dart';
-import 'package:colonizethis_world/src/utils/graph_traversal.dart';
-import '../world_constants.dart';
-
-void _tryEnqueueSeaConnectedPortExpansion({
-  required String portKey,
-  required Set<String> connected,
-  required Set<String> owned,
-  required Map<String, TileMapResult> tileMapByRegion,
-  required Map<String, int> pathCap,
-  required Queue<String> expansionSeedQueue,
-}) {
-  if (connected.contains(portKey)) return;
-  final coords = parseTileKeyCoordinates(portKey);
-  if (coords == null) return;
-  final fullProvinceId = '${coords.regionId}|${coords.provinceLocalId}';
-  if (!owned.contains(fullProvinceId)) return;
-  if (tileMapByRegion[coords.regionId] == null) return;
-  if (coords.x < 0 || coords.y < 0) return;
-
-  connected.add(portKey);
-  pathCap[portKey] = 4;
-  expansionSeedQueue.add(portKey);
-}
-
-void _removeBlockadedPortTilesExceptCapital({
-  required Set<String> connected,
-  required Map<String, int> pathCap,
-  required Set<String> blockadedPortProvinces,
-  required String capitalProvinceId,
-}) {
-  for (final key in connected.toList()) {
-    final fullProvinceId = fullProvinceIdFromTileKey(key);
-    if (fullProvinceId == null) continue;
-    if (!blockadedPortProvinces.contains(fullProvinceId)) continue;
-    if (fullProvinceId == capitalProvinceId) continue;
-    connected.remove(key);
-    pathCap.remove(key);
-  }
-}
 
 /// Resolves the connected tile set + path transport cap for a single faction's
 /// capital. SPEC/game/capital-and-connectivity. Shared by Great-Power and
@@ -98,7 +57,7 @@ ConnectivityResult connectedTilesForPlayer({
   bool expandsViaRoadOrPort(String tileKey) =>
       (tileState.roadLevel(tileKey) > 0) || portInfo.containsKey(tileKey);
 
-  _runConnectivityPropagation(
+  runConnectivityRoadPropagation(
     queue: Queue<String>()..add(capitalKey),
     connected: connected,
     pathCap: pathCap,
@@ -120,7 +79,7 @@ ConnectivityResult connectedTilesForPlayer({
         k,
   };
 
-  final seaConnectedPortKeys = _seaConnectedPortKeysForCapital(
+  final seaConnectedPortKeys = seaConnectedPortKeysForCapital(
     capital: capital,
     worldState: worldState,
     topology: topology,
@@ -136,7 +95,7 @@ ConnectivityResult connectedTilesForPlayer({
 
   final expansionSeedQueue = Queue<String>();
   for (final portKey in seaConnectedPortKeys) {
-    _tryEnqueueSeaConnectedPortExpansion(
+    tryEnqueueSeaConnectedPortExpansion(
       portKey: portKey,
       connected: connected,
       owned: owned,
@@ -146,7 +105,7 @@ ConnectivityResult connectedTilesForPlayer({
     );
   }
 
-  _runConnectivityPropagation(
+  runConnectivityRoadPropagation(
     queue: expansionSeedQueue,
     connected: connected,
     pathCap: pathCap,
@@ -161,7 +120,7 @@ ConnectivityResult connectedTilesForPlayer({
 
   // SPEC § Blockade: no tiles in a blockaded port province contribute; remove any tile in such a province (except capital province: its tiles remain when it is blockaded, only sea connectivity is severed).
   if (blockadedPortProvinces.isNotEmpty) {
-    _removeBlockadedPortTilesExceptCapital(
+    removeBlockadedPortTilesExceptCapital(
       connected: connected,
       pathCap: pathCap,
       blockadedPortProvinces: blockadedPortProvinces,
@@ -170,7 +129,7 @@ ConnectivityResult connectedTilesForPlayer({
   }
 
   final connectedByRoadRule = Set<String>.from(connected);
-  _applyTownRuleConnectivityClosure(
+  applyTownRuleConnectivityClosure(
     townByTileKey: townByTileKey,
     tileMapByRegion: tileMapByRegion,
     provinceIdsByType: provinceIdsByType,
@@ -186,188 +145,4 @@ ConnectivityResult connectedTilesForPlayer({
     pathTransportCap: pathCap,
     connectedByRoadRule: connectedByRoadRule,
   );
-}
-
-void _runConnectivityPropagation({
-  required Queue<String> queue,
-  required Set<String> connected,
-  required Map<String, int> pathCap,
-  required WorldState worldState,
-  required Map<String, (String, String)> portTileToProvinceSeaZone,
-  required Map<String, TileMapResult> tileMapByRegion,
-  required Set<String> provinceIdsByType,
-  required Set<String> ownedProvinceIds,
-  required bool Function(String tileKey) canExpandFrom,
-  ConnectivityHotPathMetrics? metrics,
-}) {
-  propagateConnectivityBottleneckQueue(
-    queue: queue,
-    connected: connected,
-    pathCap: pathCap,
-    onDequeue: metrics?.recordConnectivityBottleneckDequeue,
-    shouldExpandEdgesFrom: (key) {
-      final coords = parseTileKeyCoordinates(key);
-      if (coords == null) return false;
-      final fullProvinceId = '${coords.regionId}|${coords.provinceLocalId}';
-      if (coords.x < 0 || coords.y < 0) return false;
-      if (!ownedProvinceIds.contains(fullProvinceId)) return false;
-      final map = tileMapByRegion[coords.regionId];
-      if (map == null) return false;
-      return canExpandFrom(key);
-    },
-    neighborsOf: (key) {
-      final coords = parseTileKeyCoordinates(key);
-      if (coords == null) return const <String>[];
-      if (coords.x < 0 || coords.y < 0) return const <String>[];
-      final map = tileMapByRegion[coords.regionId];
-      if (map == null) return const <String>[];
-      return adjacentTileKeys(
-        coords.regionId,
-        coords.provinceLocalId,
-        coords.x,
-        coords.y,
-        map,
-        provinceIdsByType,
-      );
-    },
-    transportLevelAt: (neighbor) =>
-        transportLevelAtTile(worldState, neighbor, portTileToProvinceSeaZone),
-  );
-}
-
-Set<String> _seaConnectedPortKeysForCapital({
-  required CapitalTile capital,
-  required WorldState worldState,
-  required MapTopology topology,
-  required Map<String, TileMapResult> tileMapByRegion,
-  required Set<String> provinceIdsByType,
-  required Set<String> seaZoneNodeIds,
-  required Set<String> ownedProvinceIds,
-  required Set<String> blockadedPortProvinces,
-  required Set<String> capitalRegionPortKeys,
-  ConnectivityHotPathMetrics? metrics,
-}) {
-  final out = <String>{};
-  final capitalProvinceBlockaded = blockadedPortProvinces.contains(
-    capital.provinceId,
-  );
-  final capitalOnSeaboard = isCapitalTileOnSeaboard(
-    capital,
-    tileMapByRegion,
-    provinceIdsByType,
-  );
-  if (capitalOnSeaboard && !capitalProvinceBlockaded) {
-    final prefixedTopology = topologyUsesPrefixedIds(topology);
-    final provinceIdForLookup = prefixedTopology
-        ? capital.provinceId
-        : (ProvinceId.isPrefixed(capital.provinceId)
-              ? ProvinceId.localIdFrom(capital.provinceId)
-              : capital.provinceId);
-    final capitalSeaZones = seaZoneIdsAdjacentToProvince(
-      topology,
-      provinceIdForLookup,
-      regionId: ProvinceId.isPrefixed(capital.provinceId)
-          ? ProvinceId.regionIdFrom(capital.provinceId)
-          : null,
-    );
-    final seaReachable = seaZonesReachableBySeaPath(
-      topology,
-      capitalSeaZones,
-      onDequeue: metrics?.recordSeaZoneBfsDequeue,
-    );
-    for (final entry in worldState.portsByProvinceSeaboard.entries) {
-      final portMeta = decodePortSeaboardRegistryKey(entry.key);
-      if (portMeta == null) continue;
-      if (blockadedPortProvinces.contains(portMeta.fullProvinceId)) continue;
-      final seaZoneIdForReachable = prefixedTopology && portMeta.isPrefixedKey
-          ? '${portMeta.regionId}|${portMeta.seaZoneId}'
-          : portMeta.seaZoneId;
-      if (!seaReachable.contains(seaZoneIdForReachable)) continue;
-      if (!ownedProvinceIds.contains(portMeta.fullProvinceId)) continue;
-      out.add(entry.value);
-    }
-    return out;
-  }
-  if (!capitalOnSeaboard) {
-    for (final portKey in capitalRegionPortKeys) {
-      final portProvinceId = fullProvinceIdFromTileKey(portKey);
-      if (portProvinceId == null) {
-        out.add(portKey);
-        continue;
-      }
-      if (!blockadedPortProvinces.contains(portProvinceId)) {
-        out.add(portKey);
-      }
-    }
-  }
-  return out;
-}
-
-/// § Connectivity (Game Rule) Town rule: 4-adjacent to a connected town in the **same** province.
-///
-/// [townByTileKey] is the player-scoped town-tile → province map prepared by
-/// `_buildPerPlayerProvinceCaches` in a single dual-region scan (Refs #2394).
-void _applyTownRuleConnectivityClosure({
-  required Map<String, Province> townByTileKey,
-  required Map<String, TileMapResult> tileMapByRegion,
-  required Set<String> provinceIdsByType,
-  required WorldState worldState,
-  required Map<String, (String, String)> portTileToProvinceSeaZone,
-  required Set<String> connected,
-  required Map<String, int> pathCap,
-  ConnectivityHotPathMetrics? metrics,
-}) {
-  final pendingTowns = Queue<String>();
-  final queuedTowns = <String>{};
-  final expandedTowns = <String>{};
-
-  void enqueueTownForExpansion(String tk) {
-    if (!connected.contains(tk)) return;
-    assert(
-      !expandedTowns.contains(tk),
-      'town-rule worklist must not re-enqueue an already-expanded town tile: $tk',
-    );
-    if (!queuedTowns.add(tk)) return;
-    pendingTowns.add(tk);
-  }
-
-  for (final tk in townByTileKey.keys) {
-    enqueueTownForExpansion(tk);
-  }
-
-  while (pendingTowns.isNotEmpty) {
-    final tk = pendingTowns.removeFirst();
-    metrics?.recordTownRuleWorklistDequeue();
-    queuedTowns.remove(tk);
-    expandedTowns.add(tk);
-
-    final province = townByTileKey[tk];
-    if (province == null) continue;
-
-    final coords = parseTileKeyCoordinates(tk);
-    if (coords == null) continue;
-    if (coords.x < 0 || coords.y < 0) continue;
-    final map = tileMapByRegion[coords.regionId];
-    if (map == null) continue;
-
-    for (final d in kGridNeighborsCardinal4) {
-      final nx = coords.x + d.$1;
-      final ny = coords.y + d.$2;
-      if (nx < 0 || nx >= map.width || ny < 0 || ny >= map.height) continue;
-      final cell = map.cell(nx, ny);
-      if (!isLandProvinceGridCell(cell, coords.regionId, provinceIdsByType)) {
-        continue;
-      }
-      if (cell != coords.provinceLocalId) continue;
-      final nKey = CapitalTile.tileKey(coords.regionId, province.id, nx, ny);
-      if (connected.contains(nKey)) continue;
-      connected.add(nKey);
-      pathCap[nKey] =
-          pathCap[tk] ??
-          transportLevelAtTile(worldState, tk, portTileToProvinceSeaZone);
-      if (townByTileKey.containsKey(nKey)) {
-        enqueueTownForExpansion(nKey);
-      }
-    }
-  }
 }
